+++ src/sudoku_board.rs
@@ -14,11 +14,11 @@ impl SudokuBoard {
         SudokuBoard {tiles: sd_array}
     }
 
-    fn make_guess(&mut self, digit_index: u32, guess_digit: u32) {
+    fn make_guess(&mut self, digit_index: u8, guess_digit: u8) {
         self.tiles[digit_index as usize] = SudokuDigit::Guess(guess_digit);
     }
 
-    fn make_known(&mut self, digit_index: u32, known_digit: u32) {
+    fn make_known(&mut self, digit_index: u8, known_digit: u8) {
         self.tiles[digit_index as usize] = SudokuDigit::Known(known_digit)
     }
